pyproject.toml
[project]
name = "tts-project"
version = "0.1.0"
description = "Proyecto base de TTS entrenado desde cero para una voz propia."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.2",
  "gradio>=5.0.0",
  "numpy>=2.0.0",
  "pyyaml>=6.0.0",
  "soundfile>=0.12.1",
  "torch>=2.7.0",
  "torchaudio>=2.7.0",
  "uvicorn>=0.30.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
